requirements.txt
blinker==1.9.0
click==8.1.8
dbus-python==1.3.2
distro==1.9.0
dnf==4.23.0
Flask==3.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
libcomps==0.1.21
libdnf==0.74.0
MarkupSafe==3.0.2
mercurial==6.8.2
nftables==0.1
packaging==25.0
pluggy==1.5.0
PyGObject==3.48.2
pyparsing==3.1.2
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==1.1.0
rpm==4.20.1
six==1.16.0
systemd-python==235
Werkzeug==3.1.3
